# Optimize domain processing.