--- a/src/url_aggregator.cpp
+++ b/src/url_aggregator.cpp
@@ -458,23 +458,50 @@ ada_really_inline bool url_aggregator::parse_host(std::string_view input) {
   // to ASCII with domain and false. The most common case is an ASCII input, in
   // which case we do not need to call the expensive 'to_ascii' if a few
   // conditions are met: no '%' and no 'xn-' subsequence.
-  std::string _buffer = std::string(input);
-  // This next function checks that the result is ascii, but we are going to
-  // to check anyhow with is_forbidden.
-  // bool is_ascii =
-  unicode::to_lower_ascii(_buffer.data(), _buffer.size());
-  bool is_forbidden = unicode::contains_forbidden_domain_code_point(
-      _buffer.data(), _buffer.size());
-  if (is_forbidden == 0 && _buffer.find("xn-") == std::string_view::npos) {
+
+  // Often, the input does not contain any forbidden code points, and no upper
+  // case ASCII letter, then we can just copy it to the buffer. We want to
+  // optimize for such a common case.
+  uint8_t is_forbidden_or_upper =
+      unicode::contains_forbidden_domain_code_point_or_upper(input.data(),
+                                                             input.size());
+  // Minor optimization opportunity:
+  // contains_forbidden_domain_code_point_or_upper could be extend to check for
+  // the presence of characters that cannot appear in the ipv4 address and we
+  // could also check whether x and n and - are present, and so we could skip
+  // some of the checks below. However, the gains are likely to be small, and
+  // the code would be more complex.
+  if (is_forbidden_or_upper == 0 &&
+      input.find("xn-") == std::string_view::npos) {
     // fast path
-    update_base_hostname(_buffer);
+    update_base_hostname(input);
     if (checkers::is_ipv4(get_hostname())) {
       ada_log("parse_host fast path ipv4");
       return parse_ipv4(get_hostname());
     }
     ada_log("parse_host fast path ", get_hostname());
     return true;
+  } else if (is_forbidden_or_upper == 2) {
+    // We have encountered at least one upper case ASCII letter, let us
+    // try to convert it to lower case. If there is no 'xn-' in the result,
+    // we can then use a secondary fast path.
+    std::string _buffer = std::string(input);
+    unicode::to_lower_ascii(_buffer.data(), _buffer.size());
+    if (input.find("xn-") == std::string_view::npos) {
+      // secondary fast path when input is not all lower case
+      update_base_hostname(input);
+      if (checkers::is_ipv4(get_hostname())) {
+        ada_log("parse_host fast path ipv4");
+        return parse_ipv4(get_hostname());
+      }
+      ada_log("parse_host fast path ", get_hostname());
+      return true;
+    }
   }
+  // We have encountered at least one forbidden code point or the input contains
+  // 'xn-' (case insensitive), so we need to call 'to_ascii' to perform the full
+  // conversion.
+
   ada_log("parse_host calling to_ascii");
   std::optional<std::string> host = std::string(get_hostname());
   is_valid = ada::unicode::to_ascii(host, input, input.find('%'));