Currently, ada::url_aggregator will always copy the domain string to a temporary `std::string` which is the processed and finally copied on our buffer. This processing turns out to have an overhead of nearly 10% in some realistic cases.

This PR optimizes the case when the input domain is all lowercase, which is the common case. We could get smarter, but it seems to be good enough.

Using `benchdata` which defaults on @anonrig 's large URL file.

Before

```
BasicBench_AdaURL_aggregator_href             2394 ns         2390 ns       293043 speed=338.136M/s time/byte=2.95739ns time/url=217.234ns url/s=4.60334M/s
```

After


```
BasicBench_AdaURL_aggregator_href             2184 ns         2180 ns       315622 speed=370.673M/s time/byte=2.69779ns time/url=198.165ns url/s=5.0463M/s
```

So a 10% boost on this test, with an M2/LLVM14. I know how to achieve an extra 2% but it would result in more complicated code (it is described in the comments).